discovery: dont_forward bit in msg flag.
In case the __dont_forward__ is set in the ChanUpdate msg lnd
will not forward this msg to the gossiper hence broadcasting
it to all its peers. This makes sure we follow the network rules
and dont forward ChanUpdates which do not signal it.

In addition an edge case was fixed where we would remove a channel
from the db and in parallel receiving a new ChanUpdate for it. This
lead to the case of throttling messages from that peer not receiving
the new ChanUpdate (with the __dont_forward__ bit unset).