htlcswitch: change behaviour of held htlc map
In case we cannot successful resolve the HTLC we make sure we do
not remove the htlc from the held map. This fixes a bug where when
an interceptor would be required but not available during the
resolution of an HTLC we still can resolve the HTLC when the
interceptor comes back online.